requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
apoc==0.12.0
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
faiss-cpu==1.8.0.post1
fastapi==0.109.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.2
importlib-metadata==6.11.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.0
langchain-core==0.1.21
langsmith==0.0.87
langchain-community==0.0.20
langchainhub==0.1.14
langchain-openai==0.0.2
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
neo4j==5.14.1
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
openai==1.14.0
opentelemetry-api==1.22.0
orjson==3.10.6
packaging==23.2
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
protobuf==5.27.2
py==1.11.0
pyarrow==17.0.0
pyclesperanto_prototype==0.24.4
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
PyMySQL==1.1.1
pyopencl==2024.2.7
pyparsing==3.1.2
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytools==2024.1.11
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
retry==0.9.2
rich==13.7.1
rpds-py==0.19.1
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
starlette==0.35.1
streamlit==1.37.0
striprtf==0.0.26
tenacity==8.5.0
threadpoolctl==3.5.0
tifffile==2024.7.24
tiktoken==0.5.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
transforms3d==0.4.2
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.25.0
watchdog==4.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2

#Adds MS Bot functionality
#openai==1.34.0
#python-dotenv==1.0.1
#fastapi[uvicorn]==0.111.0
#pydantic==2.8.0
#redis==5.0.6
#pytz
#streamlit==1.36.0
#requests
#aiofiles
#pdf2image==1.17.0
#Pillow==10.3.0
#botbuilder-core
#botbuilder-schema
#botbuilder-integration-aiohttp